dexter_phase_3/requirements.txt
openai>=1.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
pyyaml>=6.0
openpyxl>=3.1.0
pgeocode>=0.4.0
